Fixing compilation warnings.

+++ roo_testing/frameworks/esp-idf-v4.4.1/components/esp_ringbuf/ringbuf.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
@@ -1425,10 +1426,10 @@ void xRingbufferPrintInfo(RingbufHandle_t xRingbuffer)
 {
     Ringbuffer_t *pxRingbuffer = (Ringbuffer_t *)xRingbuffer;
     configASSERT(pxRingbuffer);
-    printf("Rb size:%d\tfree: %d\trptr: %d\tfreeptr: %d\twptr: %d, aptr: %d\n",
-           pxRingbuffer->xSize, prvGetFreeSize(pxRingbuffer),
-           pxRingbuffer->pucRead - pxRingbuffer->pucHead,
-           pxRingbuffer->pucFree - pxRingbuffer->pucHead,
-           pxRingbuffer->pucWrite - pxRingbuffer->pucHead,
-           pxRingbuffer->pucAcquire - pxRingbuffer->pucHead);
+        printf("Rb size:%zu\tfree: %zu\trptr: %td\tfreeptr: %td\twptr: %td, aptr: %td\n",
+            (size_t)pxRingbuffer->xSize, (size_t)prvGetFreeSize(pxRingbuffer),
+            (ptrdiff_t)(pxRingbuffer->pucRead - pxRingbuffer->pucHead),
+            (ptrdiff_t)(pxRingbuffer->pucFree - pxRingbuffer->pucHead),
+            (ptrdiff_t)(pxRingbuffer->pucWrite - pxRingbuffer->pucHead),
+            (ptrdiff_t)(pxRingbuffer->pucAcquire - pxRingbuffer->pucHead));
 }
